# frontend-test-go-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The registry API for SwaggerHub"
  version: "1.0.97"
  title: "SwaggerHub Registry API"
  contact:
    name: "SwaggerHub"
    url: "http://swaggerhub.com"
    email: "info@swaggerhub.com"
host: "dev-api.swaggerhub.com"
...
schemes:
- "https"
produces:
- "application/json"
paths:
...
  /apis/{owner}/{api}/{version}/integrations:
    post:
      tags:
      - "Integrations"
      - "APIs"
      summary: "Creates an Integration for the given API"
      description: "Creates an Integration for the given API. The integration type\
        \ is determined by which model is used in the body."
      operationId: "createIntegration"
      consumes:
      - "application/json"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "api"
        in: "path"
        description: "API identifier"
        required: true
        type: "string"
        x-exportParamName: "Api"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      - in: "body"
        name: "integrationTypeConfiguration"
        description: "Can be one of the following objects: `AutoMockingIntegration`,\
          \ `AzureDevopsServerIntegration`, `AzureDevopsServicesIntegration`, `BitbucketCloudIntegration`,\
          \ `BitbucketServerIntegration`, `GitHubIntegration`, `GitHubEnterpriseIntegration`\
          \ or `GitLabIntegration` or `WebhookIntegration`\n"
        required: true
        schema:
          type: "object"
          example:
            configType: "GITHUB"
            name: "Java client SDK"
            token: "abcd12345"
            owner: "myOrg"
            repository: "petstore"
            branch: "develop"
            target: "java"
            syncMethod: "Advanced Sync"
            outputFolder: "sdk/java"
            providedPaths:
            - "*"
            managedPaths: []
            ignoredPaths: []
            enabled: true
        x-exportParamName: "IntegrationTypeConfiguration"
      responses:
        "201":
          description: "Integration was successfully created"
        "400":
          description: "The integration configuration was invalid"
        "401":
          description: "Unauthorized"
        "403":
          description: "Access denied"
        "404":
          description: "owner, api, version combination not found"
        "409":
          description: "Integration name already exists for API for different integration\
            \ type"
        "415":
          description: "Invalid content type"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.controllers.IntegrationsController"
  /apis/{owner}/{api}/{version}/integrations/{integrationId}:
    get:
      tags:
      - "Integrations"
      - "APIs"
      summary: "Retrieves the specified API integration."
      description: "Retrieves an integration for the given API."
      operationId: "getIntegrationById"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "api"
        in: "path"
        description: "API identifier"
        required: true
        type: "string"
        x-exportParamName: "Api"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      - name: "integrationId"
        in: "path"
        description: "integration identifier"
        required: true
        type: "string"
        x-exportParamName: "IntegrationId"
      responses:
        "200":
          description: "Can be one of the following objects: `AutoMockingIntegration`,\
            \ `AzureDevopsServerIntegration`, `AzureDevopsServicesIntegration`, `BitbucketCloudIntegration`,\
            \ `BitbucketServerIntegration`, `GitHubIntegration`, `GitHubEnterpriseIntegration`\
            \ or `GitLabIntegration` or `WebhookIntegration`\n"
          schema:
            type: "object"
            example:
              configType: "GITHUB"
              name: "Java client SDK"
              owner: "myOrg"
              repository: "petstore"
              branch: "develop"
              target: "java"
              syncMethod: "Advanced Sync"
              outputFolder: "sdk/java"
              providedPaths:
              - "*"
              managedPaths: []
              ignoredPaths: []
              enabled: true
        "401":
          description: "Unauthorized"
        "403":
          description: "User doesn't have permission to retrieve integrations in this\
            \ API"
        "404":
          description: "owner, api, version, integrationId combination not found"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.controllers.IntegrationsController"
    delete:
      tags:
      - "Integrations"
      - "APIs"
      summary: "Delete the specified API integration."
      description: "Delete an existing integration for the given API."
      operationId: "deleteIntegration"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "api"
        in: "path"
        description: "API identifier"
        required: true
        type: "string"
        x-exportParamName: "Api"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      - name: "integrationId"
        in: "path"
        description: "integration identifier"
        required: true
        type: "string"
        x-exportParamName: "IntegrationId"
      responses:
        "204":
          description: "Integration was successfully deleted"
        "401":
          description: "Unauthorized"
        "403":
          description: "User doesn't have permission to delete integrations in this\
            \ API"
        "404":
          description: "The specified API or integration ID was not found"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.controllers.IntegrationsController"
securityDefinitions:
  TokenSecured:
    description: "Specify your SwaggerHub API key found in https://app.swaggerhub.com/settings/apiKey."
    type: "apiKey"
    name: "Authorization"
    in: "header"
...